NewsBreif/requirements.txt
# Core Dependencies
pathway>=0.10.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI & NLP
google-genai>=1.0.0
google-generativeai>=0.8.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data Processing
httpx>=0.25.0
python-dateutil>=2.8.0

# Frontend
streamlit>=1.28.0

# Configuration
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

